Bug 3164: Total memory info display 32-bit overflows
Use double instead of long. This removes any Squid-caused overflow wraps.
Overflows may still appear since the underlying library is 32-bit.

+++ include/util.h
@@ -150,7 +150,7 @@ extern void gb_flush(gb_t *);  /* internal, do not use this */
 /*
  * Returns the amount of known allocated memory
  */
-int statMemoryAccounted(void);
+double statMemoryAccounted(void);
 
 /* Windows Port */
 /* win32lib.c */
